Passed all unit tests

+++ model/domain/cake.go
@@ -4,7 +4,7 @@ type Cake struct {
 	Id          int
 	Title       string
 	Description string
-	Rating      float32
+	Rating      float64
 	Image       string
 	CreatedAt   string
 	UpdatedAt   string
